alea/broadcast: simplify batch cut logic

+++ pkg/alea/alea.go
@@ -45,7 +45,7 @@ type Params struct {
 
 	// Number of batches that the broadcast component tries to have broadcast at all times in own queue
 	// Must be at least 1
-	TargetOwnUnagreedBatchCount int
+	TargetOwnUnagreedBatchCount uint64
 
 	// Time to wait before retrying batch creation
 	// Must be non-negative
